Reorganize command namings and structure
